auth tests bsky
